fix(Storage): adjust system name to properly show in search
Since Ifae5819bf1d7a1a675df1fa67bd6dcdfa074b584 the system name uses
xliff:g as a placeholder for the version number. This appears as
'Android %s' in the universal search results. There is no way
(currently?) to fill in the version number when obtaining the data for
the search from the XML resource.

As a pragmatic fix, this commit removes the version number from the
system name.

Fixes: https://gitlab.com/LineageOS/issues/android/-/issues/8308
Change-Id: If76c8f2380b9660b0ede2c00d5da6eaf1c84f868